+++ src/Libraries/SmartStore.Data/Mapping/Directory/CurrencyMap.cs
@@ -14,6 +14,7 @@ namespace SmartStore.Data.Mapping.Directory
             this.Property(c => c.DisplayLocale).HasMaxLength(50);
             this.Property(c => c.CustomFormatting).HasMaxLength(50);
             this.Property(c => c.Rate).HasPrecision(18, 4);
+			this.Property(c => c.DomainEndings).HasMaxLength(1000);
         }
     }
 }